Refactor floppy (SWIM/IWM)
